Simplify banner path resolution
Also remove `tests/support/Sandbox` in favor of reusing `MockModule`.
