+++ tests/support/MockModule.ts
@@ -1,28 +1,77 @@
 import { RootRequire } from '@dojo/interfaces/loader';
 import * as mockery from 'mockery';
-import Sandbox, { getBasePath, load, unload } from './Sandbox';
+import * as sinon from 'sinon';
 
 declare const require: RootRequire;
+const dojoNodePlugin = 'intern/dojo/node';
+
+function load(modulePath: string): any {
+	const mid = `${dojoNodePlugin}!${modulePath}`;
+	return require(mid);
+}
+
+function unload(modulePath: string): void {
+	const abs = require.toUrl(modulePath);
+	const plugin = require.toAbsMid(dojoNodePlugin);
+	require.undef(`${plugin}!${abs}`);
+}
+
+function resolvePath(basePath: string, modulePath: string): string {
+	return modulePath.replace('./', `${basePath}/`);
+}
+
+function getBasePath(modulePath: string): string {
+	const chunks = modulePath.split('/');
+	chunks.pop();
+	return chunks.join('/');
+}
 
 export default class MockModule {
+	private basePath: string;
 	private moduleUnderTestPath: string;
-	private sandbox: Sandbox;
+	private mocks: any;
+	private sandbox: sinon.SinonSandbox;
 
 	constructor(moduleUnderTestPath: string) {
+		this.basePath = getBasePath(moduleUnderTestPath);
 		this.moduleUnderTestPath = moduleUnderTestPath;
-		this.sandbox = new Sandbox(getBasePath(moduleUnderTestPath));
+		this.sandbox = sinon.sandbox.create();
+		this.mocks = {};
 	}
 
 	dependencies(dependencies: string[]): void {
-		this.sandbox.dependencies(dependencies);
+		dependencies.forEach((dependencyName) => {
+			let dependency = load(resolvePath(this.basePath, dependencyName));
+			const mock: any = {};
+
+			for (let prop in dependency) {
+				if (typeof dependency[prop] === 'function') {
+					mock[prop] = function () {};
+					this.sandbox.stub(mock, prop);
+				} else {
+					mock[prop] = dependency[prop];
+				}
+			}
+
+			if (typeof dependency === 'function') {
+				const ctor = this.sandbox.stub().returns(mock);
+				Object.assign(ctor, mock);
+				mockery.registerMock(dependencyName, ctor);
+				mock.ctor = ctor;
+			}
+			else {
+				mockery.registerMock(dependencyName, mock);
+			}
+			this.mocks[dependencyName] = mock;
+		});
 	}
 
 	getMock(dependencyName: string): any {
-		return this.sandbox.getMock(dependencyName);
+		return this.mocks[dependencyName];
 	}
 
 	getModuleUnderTest(): any {
-		this.sandbox.start();
+		this.start();
 		const allowable = require.toUrl(this.moduleUnderTestPath) + '.js';
 		mockery.registerAllowable(allowable, true);
 		return load(this.moduleUnderTestPath);
@@ -30,6 +79,12 @@ export default class MockModule {
 
 	destroy(): void {
 		unload(this.moduleUnderTestPath);
-		this.sandbox.destroy();
+		this.sandbox.restore();
+		mockery.deregisterAll();
+		mockery.disable();
+	}
+
+	start() {
+		mockery.enable({ warnOnUnregistered: false, useCleanCache: true });
 	}
 }
